Stylesheets and type definitions
