+++ package.json
@@ -13,14 +13,17 @@
   "dependencies": {},
   "devDependencies": {
     "@types/dompurify": "^2.0.2",
+    "@types/fs-extra": "^9.0.1",
     "@types/jsdom": "^16.2.3",
     "@types/live-server": "^1.2.0",
     "@types/marked": "^1.1.0",
     "date-fns": "^2.14.0",
     "dompurify": "^2.0.12",
     "front-matter": "^4.0.2",
+    "fs-extra": "^9.0.1",
     "gh-pages": "^3.1.0",
     "gray-matter": "^4.0.2",
+    "jest": "^26.1.0",
     "js-beautify": "^1.11.0",
     "jsdom": "^16.2.2",
     "live-server": "^1.2.1",
