+++ public/assets/css/style.css
@@ -1,35 +1,84 @@
 @font-face {
   font-family: "Manrope";
-  src: url("../fonts/manrope/Manrope-ExtraLight.ttf") format("truetype");
+  src: url("assets/fonts/manrope/Manrope-ExtraLight.ttf") format("truetype");
   font-weight: 100;
 }
 @font-face {
   font-family: "Manrope";
-  src: url("../assets/fonts/manrope/Manrope-Light.ttf") format("truetype");
+  src: url("assets/fonts/manrope/Manrope-Light.ttf") format("truetype");
   font-weight: 200;
 }
 @font-face {
   font-family: "Manrope";
-  src: url("../assets/fonts/manrope/Manrope-Regular.ttf") format("truetype");
+  src: url("assets/fonts/manrope/Manrope-Regular.ttf") format("truetype");
   font-weight: 400;
 }
 @font-face {
   font-family: "Manrope";
-  src: url("../assets/fonts/manrope/Manrope-SemiBold.ttf") format("truetype");
+  src: url("assets/fonts/manrope/Manrope-SemiBold.ttf") format("truetype");
   font-weight: 500;
 }
 @font-face {
   font-family: "Manrope";
-  src: url("../assets/fonts/manrope/Manrope-Bold.ttf") format("truetype");
+  src: url("assets/fonts/manrope/Manrope-Bold.ttf") format("truetype");
   font-weight: 600;
 }
 @font-face {
   font-family: "Manrope";
-  src: url("../assets/fonts/manrope/Manrope-ExtraBold.ttf") format("truetype");
+  src: url("assets/fonts/manrope/Manrope-ExtraBold.ttf") format("truetype");
   font-weight: 900;
 }
 @font-face {
   font-family: "Manrope";
-  src: url("../assets/fonts/manrope/Manrope-VariableFont_wght.ttf") format("truetype");
+  src: url("assets/fonts/manrope/Manrope-VariableFont_wght.ttf") format("truetype");
   font-weight: 200 800;
+}
+
+
+body {
+  margin: 4rem 10rem 4rem 4rem; 
+  padding: 0;
+  cursor: default;
+}
+
+h1, h2, h3, h4, h5, h6, p, span, a {
+  font-family: "Manrope", sans-serif;
+  color: rgb(90,90,90);
+  font-weight: 475;
+  font-size: 72px;
+  display: inline-block;
+  padding: 0;
+  margin: 0; 
+  orphans: 0;
+  widows: 0;
+}
+
+h1 {
+  color: rgb(0,0,0);
+}
+
+a, a:visited, a:active, a:hover {
+  cursor: pointer;
+}
+
+.fade-enter-active,
+.fade-leave-active {
+  transition-duration: 250ms;
+  transition-property: opacity;
+  transition-timing-function: ease-in-out;
+}
+
+.fade-enter,
+.fade-leave-active {
+  opacity: 0
+}
+
+.debug-link-active {
+  text-decoration: none;
+  padding: 0; 
+  margin: 0 1rem 0 0;
+}
+
+.tba {
+  cursor: not-allowed;
 }
