+++ src/templates/index.html.ts
@@ -5,15 +5,15 @@ export const template = (meta:any, markup:string) => {
 		<!doctype html>
 		<html>
 			<head>
-				<title>${meta.title} - Leslie Owusu-Appiah</title>
+				<title>${meta.title}</title>
 				<meta name="keywords" content="" />
 				<meta name="" content="" />
 				<meta name="" content="" />
-				<link rel="stylesheet" href="/css/style.css" />
+				<link rel="stylesheet" href="/assets/css/style.css" />
 			</head>
 			<body data-template="${meta.template}">
 				<header>
-					<h1>127.0.0.1</h1>
+					<h1>${meta.title}</h1>
 				</header>
 				<h1>${meta.title}</h1>
 				<small>${meta.date}</small>
